requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
